# rewrite Twine to improve performance of ToScala by 560%