--- a/build.mill
+++ b/build.mill
@@ -43,8 +43,6 @@ object `package` extends RootModule with ScalaModule {
   }
 
   object test extends ScalaTests with TestModule.ScalaTest {
-    def forkArgs = Seq("-Dline.separator=\r\n")
-
     override def ammoniteVersion = ammonite
     def ivyDeps = Agg(scalaTests, scalaCheck, javaTests)
     def sources = Task.Sources {