--- a/src/main/scala/ir/dsl/ProgramToScala.scala
+++ b/src/main/scala/ir/dsl/ProgramToScala.scala
@@ -2,7 +2,8 @@ package ir.dsl
 
 import ir.*
 import ir.dsl.{ToScala, ToScalaLines}
-import util.{Twine, indentNested, intersperse}
+import util.{intersperse}
+import util.twine.*
 
 import collection.immutable.{ListMap, SortedMap}
 import collection.immutable.{LazyList}
@@ -113,7 +114,7 @@ trait BasilIRToScala {
   def blockToScala(x: Block): Twine = {
     // XXX: using a Seq here allows the size to be known, avoiding excessive splitting.
     val commands = x.statements ++ Seq(x.jump)
-    indentNested(s"block(${x.label.toScala}", commandListToScala(commands), ")", headSep = true)
+    Twine.indentNested(s"block(${x.label.toScala}", commandListToScala(commands), ")", headSep = true)
   }
 
   def procedureToScala(x: Procedure): Twine = {
@@ -128,16 +129,16 @@ trait BasilIRToScala {
       LazyList(formalParamsToScala(x.formalInParam), formalParamsToScala(x.formalOutParam))
     }
 
-    indentNested(s"proc(${x.name.toScala}", params #::: x.blocks.to(LazyList).map(blockToScala), ")", headSep = true)
+    Twine.indentNested(s"proc(${x.name.toScala}", params #::: x.blocks.to(LazyList).map(blockToScala), ")", headSep = true)
   }
 
   def initialMemoryToScala(x: Program): Option[Twine] = None
 
   def programToScala(x: Program): Twine = {
     val main = x.mainProcedure
-    val others = x.procedures.to(LazyList).filter(_ ne main)
+    val others = x.procedures.filter(_ ne main)
     val mem = initialMemoryToScala(x)
-    indentNested("prog(", mem ++: (main #:: others).map(procedureToScala), ")")
+    Twine.indentNested("prog(", mem ++: (main +: others).map(procedureToScala), ")")
   }
 }
 
@@ -250,10 +251,10 @@ trait ToScalaWithSplitting extends BasilIRToScala {
       val chunks = cmds.grouped(10).toList
       val chunkNames = chunks.map(chunk => {
         val name = nextChunk
-        addDecl(name, indentNested("Vector(", super.commandListToScala(chunk), ")"))
+        addDecl(name, Twine.indentNested("Vector(", super.commandListToScala(chunk), ")"))
         name
       })
-      Iterable(indentNested("Vector(", chunkNames.map(LazyList(_)), ").flatten : _*"))
+      Iterable(Twine.indentNested("Vector(", chunkNames.map(Str.apply), ").flatten : _*"))
     }
 
   /**
@@ -262,11 +263,11 @@ trait ToScalaWithSplitting extends BasilIRToScala {
   override def blockToScala(x: Block): Twine =
     val size = x.statements.knownSize
     if (size >= 0 && size <= 1) {
-      super.blockToScala(x).force
+      super.blockToScala(x)
     } else {
       val name = s"`block:${x.parent.name}.${x.label}`"
-      addDecl(name, super.blockToScala(x).force)
-      LazyList(name)
+      addDecl(name, super.blockToScala(x))
+      Str(name)
     }
 
   /**
@@ -274,11 +275,11 @@ trait ToScalaWithSplitting extends BasilIRToScala {
    */
   override def procedureToScala(x: Procedure): Twine =
     if (x.blocks.isEmpty) {
-      super.procedureToScala(x).force
+      super.procedureToScala(x)
     } else {
       val name = s"`procedure:${x.name}`"
-      addDecl(name, super.procedureToScala(x).force)
-      LazyList(name)
+      addDecl(name, super.procedureToScala(x))
+      Str(name)
     }
 
   /**
@@ -289,10 +290,10 @@ trait ToScalaWithSplitting extends BasilIRToScala {
   protected def toScalaAndDeclsWith[T](f: T => Twine)(name: String, x: T): Twine =
     assert(decls.isEmpty, "repeated use of the same ToScalaWithSplitting instance is not allowed")
 
-    addDecl(name, f(x).force)
+    addDecl(name, f(x))
 
     // NOTE: scala compiler will error on duplicated names
-    indentNested("{", declsToScala(decls) ++ Iterable(LazyList(name)), "}", sep = "\n")
+    Twine.indentNested("{", declsToScala(decls) ++ Iterable(Str(name)), "}", sep = "\n")
 
   def declsToScala(decls: Map[String, Twine]): Iterable[Twine] =
     decls.map((k, v) => s"def $k = " +: v)
@@ -325,7 +326,7 @@ object ToScalaWithInitialMemory {
  */
 trait ToScalaWithInitialMemory extends BasilIRToScala {
   override def initialMemoryToScala(x: Program) =
-    Some(indentNested("Seq(", x.initialMemory.values.map(_.toScalaLines), ")"))
+    Some(Twine.indentNested("Seq(", x.initialMemory.values.map(_.toScalaLines), ")"))
 }
 
 given ToScalaLines[MemorySection] with
@@ -335,19 +336,19 @@ given ToScalaLines[MemorySection] with
         x.bytes
           .map(x => f"${x.value}%#04x")
           .grouped(32)
-          .map(x => LazyList(x.mkString(",")))
+          .map(x => Str(x.mkString(",")))
           .toSeq
 
-      val byteTwine = indentNested("Seq(", byteLines, ").map(BitVecLiteral(_, 8)).toSeq")
+      val byteTwine = Twine.indentNested("Seq(", byteLines, ").map(BitVecLiteral(_, 8)).toSeq")
 
-      // in the second argument of indentNested, list elements will be separated by newlines.
-      indentNested(
+      // in the second argument of Twine.indentNested, list elements will be separated by newlines.
+      Twine.indentNested(
         "MemorySection(",
         List(
-          LazyList(x.name.toScala, ", ", x.address.toScala, ", ", x.size.toScala),
+          Concat(List(x.name.toScala, ", ", x.address.toScala, ", ", x.size.toScala).map(Str(_))),
           byteTwine,
-          "readOnly = " #:: x.readOnly.toScalaLines,
-          "region = " #:: None.toScalaLines
+          "readOnly = " +: x.readOnly.toScalaLines,
+          "region = " +: None.toScalaLines
         ),
         ")"
       )