--- a/src/main/scala/ir/dsl/ProgramToScala.scala
+++ b/src/main/scala/ir/dsl/ProgramToScala.scala
@@ -298,7 +298,7 @@ trait ToScalaWithSplitting extends BasilIRToScala {
     addDecl(name, f(x))
 
     // NOTE: scala compiler will error on duplicated names
-    Twine.indentNested("{", declsToScala(decls) ++ Iterable(Twine(name)), "}", sep = "\n")
+    Twine.indentNested("{", (declsToScala(decls) ++: List(Twine(name))).intersperse(Twine("")), "}", sep = "")
 
   def declsToScala(decls: Map[String, Twine]): Iterable[Twine] =
     decls.map((k, v) => s"def $k = " +: v)