--- a/src/main/scala/ir/dsl/StatementToScala.scala
+++ b/src/main/scala/ir/dsl/StatementToScala.scala
@@ -111,11 +111,7 @@ private object CaseIR {
 
         Twine.indentNested(
           s"directCall(",
-          List(
-            outs.toSeq.toScalaLines,
-            tgt.name.toScalaLines,
-            actuals.toSeq.toScalaLines
-          ),
+          List(outs.toSeq.toScalaLines, tgt.name.toScalaLines, actuals.toSeq.toScalaLines),
           ")"
         )
       }