--- a/src/main/scala/util/RunUtils.scala
+++ b/src/main/scala/util/RunUtils.scala
@@ -963,7 +963,6 @@ object RunUtils {
       writeToFile(pp_prog(ctx.program), s"$s-output.il")
       val a = ctx.program.toScalaLines
       writeToFile(a.mkString, s"$s-output.scala")
-      println(a)
     })
     Logger.info("[!] Translating to Boogie")