… src/main/scala/util/StringUtils.scala
… src/test/scala/ir/ToScalaTest.scala