This replaces the `Twine = LazyList[String]` with a new custom Twine ADT.
```scala
case class Indent(tw: Twine) extends Twine
case class Str(s: String) extends Twine
case class Lines(ss: Iterable[Twine]) extends Twine
case class Concat(tws: Iterable[Twine]) extends Twine
```
This lets us avoid costly string manipulations and defers the construction of the string until it is needed.

The performance improvement is shown below, with the important lines being the toScalaLines and mkString.
toScalaLines is the time taken to convert the Basil IR to the Twine (either the old LazyList or the new ADT),
and mkString the time to convert the Twine to a string. With LazyList, mkString was just
concatenating the elements, but with the new ADT, the mkString function performs all of the string
indentation and formatting.

```bash
# NEW
$ ./mill runMain ir.parsing.ParseBasilIL ~/Downloads/cnflm-output.il | grep -i timer
=============== runMain ir.parsing.ParseBasilIL .../rina/Downloads/cnflm-output.il ===============
[228/228] runMain
[228] [DEBUG]   PerformanceTimer ParseBasilIL [lexed]: 16ms [checkPoint@PerformanceTimer.scala:34]
[228] [DEBUG]   PerformanceTimer ParseBasilIL [parsed]: 1084ms [checkPoint@PerformanceTimer.scala:34]
[228] [DEBUG]   PerformanceTimer ParseBasilIL [early visitor]: 119ms [checkPoint@PerformanceTimer.scala:34]
[228] [DEBUG]   PerformanceTimer ParseBasilIL [main visitor]: 406ms [checkPoint@PerformanceTimer.scala:34]
[228] [DEBUG]   PerformanceTimer ParseBasilIL [dsl resolving]: 26237ms [checkPoint@PerformanceTimer.scala:34]
[228] [DEBUG]   PerformanceTimer ParseBasilIL [toScalaLines]: 762ms [checkPoint@PerformanceTimer.scala:34]
[228] [DEBUG]   PerformanceTimer ParseBasilIL [mkString]: 197ms [checkPoint@PerformanceTimer.scala:34]
[228] [DEBUG]   PerformanceTimer ParseBasilIL [write]: 28ms [checkPoint@PerformanceTimer.scala:34]
[228/228] =============== runMain ir.parsing.ParseB...ownloads/cnflm-output.il =============== 33s
$ git co main
Switched to branch 'main'
Your branch is up to date with 'origin/main'.
$ git cherry-pick 6c85bfde3
[main 01bd41118] performance timer in ir.parsing.ParseBasilIL
 Date: Thu Jun 12 11:49:29 2025 +1000
 1 file changed, 20 insertions(+), 4 deletions(-)

# OLD
$ ./mill runMain ir.parsing.ParseBasilIL ~/Downloads/cnflm-output.il | grep -i timer
=============== runMain ir.parsing.ParseBasilIL .../rina/Downloads/cnflm-output.il ===============
[228/228] runMain
[228] [DEBUG]   PerformanceTimer ParseBasilIL [lexed]: 12ms [checkPoint@PerformanceTimer.scala:34]
[228] [DEBUG]   PerformanceTimer ParseBasilIL [parsed]: 1181ms [checkPoint@PerformanceTimer.scala:34]
[228] [DEBUG]   PerformanceTimer ParseBasilIL [early visitor]: 145ms [checkPoint@PerformanceTimer.scala:34]
[228] [DEBUG]   PerformanceTimer ParseBasilIL [main visitor]: 401ms [checkPoint@PerformanceTimer.scala:34]
[228] [DEBUG]   PerformanceTimer ParseBasilIL [dsl resolving]: 26799ms [checkPoint@PerformanceTimer.scala:34]
[228] [DEBUG]   PerformanceTimer ParseBasilIL [toScalaLines]: 5057ms [checkPoint@PerformanceTimer.scala:34]
[228] [DEBUG]   PerformanceTimer ParseBasilIL [mkString]: 1199ms [checkPoint@PerformanceTimer.scala:34]
[228] [DEBUG]   PerformanceTimer ParseBasilIL [write]: 24ms [checkPoint@PerformanceTimer.scala:34]
[228/228] =============== runMain ir.parsing.ParseB...ownloads/cnflm-output.il =============== 44s
```

**todo**
- [ ] remove performance timer doing writing things, probably.